feat: :pencil2: Add sermon comparison to Augustan's City of God in my new 'thoughts' folder.
Added a thoughts folder. Then added the sermon comparison page to include a detailed analysis of the similarities between Pastor Peter's sermon and Augustine of Hippo's work "The City of God". The page now provides a comparison of primary allegiance to Christ, engagement in earthly politics, moral and spiritual conduct, and the emphasis on supernatural change over political solutions. The changes align with Augustinian thought on dual citizenship and highlight the importance of prioritizing spiritual obligations over earthly ones.